% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{copy.to.clipboard}
\alias{copy.to.clipboard}
\title{copy.to.clipboard Function}
\usage{
copy.to.clipboard(d, getRownames = F, ...)
}
\arguments{
\item{d}{vector or data.table or data.frame or vector}

\item{getRownames}{If you want to display also the rownames, set to T}

\item{...}{all other parameters are passed to the write.table function}
}
\value{
Copies to memory
}
\description{
copy.to.clipboard Function
}
