% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{copy_all_gdx_symbols_tolist}
\alias{copy_all_gdx_symbols_tolist}
\title{copy_all_gdx_symbols_tolist Function}
\usage{
copy_all_gdx_symbols_tolist(gdx_path, exclude_symbols = c())
}
\arguments{
\item{gdx_path}{The path to the gdx file}
}
\value{
a GDX.list with all the symbols in GDX.list
}
\description{
Reads all symbols from a gdx and returns a GDX.list
}
