% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{copy_gdx_symbol_tolist_element}
\alias{copy_gdx_symbol_tolist_element}
\title{copy_gdx_symbol_tolist_element Function}
\usage{
copy_gdx_symbol_tolist_element(symbol, gdx_path)
}
\arguments{
\item{symbol}{The symbol to copy}

\item{gdx_path}{The path to the gdx file}
}
\value{
a GDX.list with the following elements, { domains,ts,data}
}
\description{
Reads a symbol from a gdx and creates a list element suitable for the GDXlist of ifmcap
}
