% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{create_map_fill}
\alias{create_map_fill}
\title{################################################################################################################################################33
create NUTS fill map}
\usage{
create_map_fill(
  DT,
  title = "Untitled",
  legend.title = "Untitled",
  legend.position.cur = "right",
  resol,
  scale_type,
  scale_fill_discrete = list(n = 3, palette = "Inferno", rev = F, alpha = 0.7),
  scale_fill_continuous = list(name = " ", low = "#FFFFFF", high = "#FF4200", mid =
    "#FFFF00", midpoint = 0, label_decimals = 1, barwidth = 15),
  nuts_year = 2016,
  external_data = "U:/SCIENTIFIC/FARM @ U/30-Projects/01-IFM-CAP/04-Model External Data",
  add.values.as.labels = F,
  values.as.labels.format = "\%0.1f",
  do.print = F,
  bg_fill = "grey80",
  xlim.cur = c(2200000, 7150000),
  ylim.cur = c(1380000, 5500000),
  debug = F
)
}
\arguments{
\item{DT}{A data.table with only two columns: \{NUTS0,NUTS1,NUTS2,NUTS3\} and value.}

\item{resol}{The NUTS resolution that the data refers to. Can be NUTS0,NUTS1,NUTS2,NUTS3}

\item{scale_type}{Should it be discrete or continous}

\item{scale_fill_discrete}{A list with options for scale_fill_manual. hcl.pals()}

\item{scale_fill_continuous}{A list with options for scale_fill_gradient}
}
\value{
a ggplot object
}
\description{
This function takes a numeric vector as input and computes various statistical measures, including mean, median, mode, and standard deviation.
}
\seealso{
\url{https://r-charts.com/spatial/choropleth-map-ggplot2/}
}
