% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fadn.R
\name{get_FARM_INFO}
\alias{get_FARM_INFO}
\title{################################################################################################################################################33
get_farm_info Function}
\usage{
get_FARM_INFO(
  database_dir = "E:/IFM_CAP2/Database2020",
  d_fadn_gdx = "d.fadn/ifm_cap_out/d_fadn_ifm_cap_data_2020.gdx",
  d_fadn_tables = "d.fadn/ifm_cap_out/d_fadn_ifm_cap_data_2020.FADN_tables.GDX",
  fadn_dictionary =
    "U:/SCIENTIFIC/FARM @ U/30-Projects/01-IFM-CAP/04-Model External Data/fadn/dictionary.xlsx",
  additional_AB_columns = c()
)
}
\arguments{
\item{database_dir}{The database directory}

\item{d_fadn_gdx}{The gdx with the export of data from the d.fadn}

\item{d_fadn_tables}{The gdx with the FADN tables}

\item{fadn_dictionary}{An excel file that has the following sheets TF14, ESC that give the correspondence of TF14 to TF14.text and ESC to ESC6.text}

\item{additional_AB_columns}{A vector with names of the TABLE_AB table to include}
}
\value{
DT with the FDALL and the NUTS0,NUTS1,NUTS2,TF14,TF8,ECS,ORGANIC,UAA,Weight and their text description
export
}
\description{
This function retrieves information data for farms from FADN, for the baseyear
}
