% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fadn.R
\name{get_FARM_INFO_allYears}
\alias{get_FARM_INFO_allYears}
\title{################################################################################################################################################33
get_FARM_INFO_allYears Function}
\usage{
get_FARM_INFO_allYears(
  d_fadn_folder = "E:/IFM_CAP2/Database2020/d.fadn/filtered_load",
  fadn_dictionary =
    "U:/SCIENTIFIC/FARM @ U/30-Projects/01-IFM-CAP/04-Model External Data/fadn/dictionary.xlsx"
)
}
\arguments{
\item{d_fadn_folder}{The folder with the data from the d.fadn that contains the rds files with the fadn tables (tableAB.rds, tableSE.rds, etc.)}

\item{fadn_dictionary}{An excel file that has the following sheets TF14, ESC that give the correspondence of TF14 to TF14.text and ESC to ESC6.text}
}
\value{
DT with the FDALL and YEAR, and the NUTS0,NUTS1,NUTS2,TF14,TF8,ECS,ORGANIC,UAA,Weight and their text description
}
\description{
This function retrieves information data for farms from FADN, for all years. It uses the d.fadn module.
}
