% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{kable_print}
\alias{kable_print}
\title{kable_print Function}
\usage{
kable_print(d, big.mark.cur = ",", digits.cur = 0, caption = "")
}
\arguments{
\item{big.mark.cur}{What character to use in order to separate thousands in digits (default ",")}

\item{digits.cur}{How many decimal digits to show (default 0)}

\item{caption}{The title to display at the top of the table}
}
\value{
Nothing
}
\description{
kable_print Function
}
