% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_with_flat_elems_to_DT}
\alias{list_with_flat_elems_to_DT}
\title{list_with_flat_elems_to_DT
This function convert a list to a data.table with two columns}
\usage{
list_with_flat_elems_to_DT(list.cur, dt.names)
}
\arguments{
\item{list.cur}{The list to convert}

\item{dt.names}{A vector with the names of the returned DT}
}
\value{
a DT
#example
TODO Provide an example of usage of this function
}
\description{
list_with_flat_elems_to_DT
This function convert a list to a data.table with two columns
}
