% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_with_named_elems_to_DT}
\alias{list_with_named_elems_to_DT}
\title{list_with_named_elems_to_DT
This function convert a list that contains named vectors to a data.table with columns the names of the vectors}
\usage{
list_with_named_elems_to_DT(list.cur, id.col.cur)
}
\arguments{
\item{list.cur}{The list to convert}

\item{id.col.cur}{How to name the names of the list}
}
\value{
a DT
example
TODO: Provide an example of the usage of this function
}
\description{
list_with_named_elems_to_DT
This function convert a list that contains named vectors to a data.table with columns the names of the vectors
}
