% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{read_from_gdx2}
\alias{read_from_gdx2}
\title{Read a symbol (set or parameter) from a gdx}
\usage{
read_from_gdx2(
  symbol,
  gdx_path,
  filters = list(),
  custom_colnames = c(),
  value_colname = "value",
  get_te = FALSE,
  get_ts = FALSE
)
}
\arguments{
\item{symbol}{A string with the symbol to load}

\item{gdx_path}{A string with the path of the GDX file}

\item{filters}{A named list specifying which elements to retain.
Each name in the list corresponds to a column name, and its value
is a character vector of the elements to keep. Filtering is case-insensitive.

\strong{Examples:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{filters = list(
  variable = c("v_Objective"),
  SCEN  = c("BSL", "y2025")
)

filters=list(variable=c("Solver","v_Objective"))

filters=list(FDALL=c("80005990095207"),variable=c("Solver","v_Objective"))
}\if{html}{\out{</div>}}}

\item{custom_colnames}{A vector for setting explicitly the names of the columns of the GDX.
Provide names only for the dimensions. The column that contains the value will take the
name \code{"value"}, unless defined otherwise via the \code{value_colname} argument.

\strong{Example:}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{custom_colnames = c("FDALL","SCEN","variable","dim1","dim2","dim3","dim4")
}\if{html}{\out{</div>}}}

\item{value_colname}{For parameters, a custom column name for the value. Default: "value"}

\item{get_te}{Retrieve the explanatory text of the set elements. It works only for set elements;
if you set it to TRUE for parameters, it will be ignored.}

\item{get_ts}{Retrieve the explanatory text of the symbol. It places it in the ts attribute}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table}} object containing the symbol.
}
\description{
This function reads a symbol from a gdx file.
}
\details{
Improvements in comparison with the gdxrrw native package:
\itemize{
\item There is no need to define if it is a parameter or a set
\item It is case insensitive (p_PMP_Linear equal to p_pmp_Linear)
\item It always handle the returned column names well
}
}
