% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{read_from_many_gdxs}
\alias{read_from_many_gdxs}
\title{Load a symbol (set or parameter) from many gdx files}
\usage{
read_from_many_gdxs(
  symbol,
  gdxfile_patt,
  custom_colnames = c(),
  filters = list(),
  keep_filename = FALSE
)
}
\arguments{
\item{symbol}{The symbol to load.}

\item{gdxfile_patt}{Character vector of file name patterns to match GDX files.
Patterns are interpreted as shell-style wildcards (as in \code{Sys.glob()}):
\itemize{
\item \code{*} matches any number of characters
\item \code{?} matches exactly one character
\item \code{[...] } matches any of the specified characters
}
Multiple patterns can be provided in a vector, e.g.
\code{c(".../batch_AT*.gdx", ".../batch_BE*.gdx")}.}

\item{custom_colnames}{Optional character vector specifying column names
to explicitly assign to the GDX data.}

\item{filters}{A named list specifying filters to apply on the loaded data.
Each element must be of the form \code{list(column_name = c("val1", "val2", ...))}.
Filtering is case-insensitive.}

\item{keep_filename}{Logical. If \code{TRUE}, an extra column \code{src_file}
is added indicating the filename that each record comes from.}
}
\value{
A list with two elements: LOADING_INFO and DATA.
}
\description{
Function to load many GDX files into one \code{data.table}.
}
\details{
The GDX files should have the same structure.
\if{html}{\out{<TODO>}} If not, an error should be thrown.
}
