% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_to_excel}
\alias{save_to_excel}
\title{Save an EXCEL.list into an excel file}
\usage{
save_to_excel(EXCEL.list.cur, save_xl_path, split.dist = 4, overwrite.cur = T)
}
\arguments{
\item{EXCEL.list.cur}{A list of elements to save in the excel file with the following structure}

\item{save_xl_path}{A string with the path of the Excel file}

\item{split.dist}{How many cells to leave between the data.frames to be saved}

\item{overwrite.cur}{T for overwrite the file if it exists}
}
\value{
NULL
#example
Provide an example of its usage
}
\description{
This function takes a numeric vector as input and computes various statistical measures, including mean, median, mode, and standard deviation.
}
\details{
TODO explain what is an EXCEL.list
}
