% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{save_to_gdx}
\alias{save_to_gdx}
\title{Saves to gdx a GDX.list}
\usage{
save_to_gdx(GDX.list.cur, save_gdx_path)
}
\arguments{
\item{GDX.list.cur}{A list of elements to save in the GDX with the following structure}

\item{save_gdx_path}{A string with the path of the GDX file}
}
\description{
TODO explain what a GDX.list is
}
