% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_wrap_factor}
\alias{str_wrap_factor}
\title{str_wrap_factor Function}
\usage{
str_wrap_factor(x, width.cur = 10, ...)
}
\value{
the factor with the labels text wrapped
}
\description{
This function takes a factor and applies str_wrap,
}
